Move Registry to the top level exports
