+++ packages/protobuf/src/codegenv1/enum.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import type { DescEnum, DescFile } from "../desc-types.js";
+import type { DescEnum, DescFile } from "../descriptors.js";
 import type { GenDescEnum } from "./types.js";
 
 /**
