+++ packages/protobuf/src/fields.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import type { MessageShape } from "./types.js";
-import type { DescField, DescMessage } from "./desc-types.js";
+import type { DescField, DescMessage } from "./descriptors.js";
 import { unsafeClear, unsafeIsSet } from "./reflect/unsafe.js";
 
 /**
