+++ packages/protobuf/src/reflect/index.ts
@@ -17,6 +17,6 @@ export * from "./names.js";
 export * from "./nested-types.js";
 export * from "./reflect.js";
 export * from "./reflect-types.js";
-export * from "./registry.js";
+export * from "../registry.js";
 export * from "./scalar.js";
 export { isReflectList, isReflectMap, isReflectMessage } from "./guard.js";
