+++ packages/protobuf/src/reflect/scalar.ts
@@ -14,73 +14,7 @@
 
 import { protoInt64 } from "../proto-int64.js";
 import { getTextEncoding } from "../wire/text-encoding.js";
-
-/**
- * Scalar value types. This is a subset of field types declared by protobuf
- * enum google.protobuf.FieldDescriptorProto.Type The types GROUP and MESSAGE
- * are omitted, but the numerical values are identical.
- */
-export enum ScalarType {
-  // 0 is reserved for errors.
-  // Order is weird for historical reasons.
-  DOUBLE = 1,
-  FLOAT = 2,
-  // Not ZigZag encoded.  Negative numbers take 10 bytes.  Use TYPE_SINT64 if
-  // negative values are likely.
-  INT64 = 3,
-  UINT64 = 4,
-  // Not ZigZag encoded.  Negative numbers take 10 bytes.  Use TYPE_SINT32 if
-  // negative values are likely.
-  INT32 = 5,
-  FIXED64 = 6,
-  FIXED32 = 7,
-  BOOL = 8,
-  STRING = 9,
-  // Tag-delimited aggregate.
-  // Group type is deprecated and not supported in proto3. However, Proto3
-  // implementations should still be able to parse the group wire format and
-  // treat group fields as unknown fields.
-  // TYPE_GROUP = 10,
-  // TYPE_MESSAGE = 11,  // Length-delimited aggregate.
-
-  // New in version 2.
-  BYTES = 12,
-  UINT32 = 13,
-  // TYPE_ENUM = 14,
-  SFIXED32 = 15,
-  SFIXED64 = 16,
-  SINT32 = 17, // Uses ZigZag encoding.
-  SINT64 = 18, // Uses ZigZag encoding.
-}
-
-/**
- * JavaScript representation of fields with 64 bit integral types (int64, uint64,
- * sint64, fixed64, sfixed64).
- *
- * This is a subset of google.protobuf.FieldOptions.JSType, which defines JS_NORMAL,
- * JS_STRING, and JS_NUMBER. Protobuf-ES uses BigInt by default, but will use
- * String if `[jstype = JS_STRING]` is specified.
- *
- * ```protobuf
- * uint64 field_a = 1; // BigInt
- * uint64 field_b = 2 [jstype = JS_NORMAL]; // BigInt
- * uint64 field_b = 2 [jstype = JS_NUMBER]; // BigInt
- * uint64 field_b = 2 [jstype = JS_STRING]; // String
- * ```
- */
-export enum LongType {
-  /**
-   * Use JavaScript BigInt.
-   */
-  BIGINT = 0,
-
-  /**
-   * Use JavaScript String.
-   *
-   * Field option `[jstype = JS_STRING]`.
-   */
-  STRING = 1,
-}
+import { LongType, ScalarType } from "../descriptors.js";
 
 /**
  * ScalarValue maps from a scalar field type to a TypeScript value type.
@@ -92,7 +26,6 @@ export type ScalarValue<
 > =
     T extends ScalarType.STRING   ? string
   : T extends ScalarType.INT32    ? number
-  : T extends ScalarType.UINT32   ? number
   : T extends ScalarType.UINT32   ? number
   : T extends ScalarType.SINT32   ? number
   : T extends ScalarType.FIXED32  ? number
@@ -159,28 +92,28 @@ export function scalarZeroValue<T extends ScalarType, L extends LongType>(
   longType: L,
 ): ScalarValue<T, L> {
   switch (type) {
+    case ScalarType.STRING:
+      return "" as ScalarValue<T>;
     case ScalarType.BOOL:
       return false as ScalarValue<T>;
-    case ScalarType.UINT64:
-    case ScalarType.FIXED64:
+    default:
+      // Handles INT32, UINT32, SINT32, FIXED32, SFIXED32.
+      // We do not use individual cases to save a few bytes code size.
+      return 0 as ScalarValue<T>;
+    case ScalarType.DOUBLE:
+    case ScalarType.FLOAT:
+      return 0.0 as ScalarValue<T>;
     case ScalarType.INT64:
+    case ScalarType.UINT64:
     case ScalarType.SFIXED64:
+    case ScalarType.FIXED64:
     case ScalarType.SINT64:
       return ((longType as number) == 0 ? protoInt64.zero : "0") as ScalarValue<
         T,
         L
       >;
-    case ScalarType.DOUBLE:
-    case ScalarType.FLOAT:
-      return 0.0 as ScalarValue<T>;
     case ScalarType.BYTES:
       return new Uint8Array(0) as ScalarValue<T>;
-    case ScalarType.STRING:
-      return "" as ScalarValue<T>;
-    default:
-      // Handles INT32, UINT32, SINT32, FIXED32, SFIXED32.
-      // We do not use individual cases to save a few bytes code size.
-      return 0 as ScalarValue<T>;
   }
 }
